Rework of Chapter1Fig1.cs
Gutted most of the script and replaced it with x and y varients mainly. Made the camera's edges as the boundry and forced the camera to Orthographic viewing (which also eliminates the need to find the depth of the scene). I made the hit boarder code easier to understand and easier to type. I could not figure out how to implement <Transform>.TranslatePoint into the script, so I stuck with <Transform>.Translate(). Script was tested and worked as intended in the demo scene.