Removed undeeded Prefabs and Scripts. Updated Chapter1Fig7.cs
Removed redundent and older scripts.
The new script functions similar to the original script. Everything is fit into one file so that there isn't as much that has to happen. I added in Time.deltaTime with the position update to help stablilize the speed to match x meters per second instead of x meters per frame. Frame looping was added so if the object goes off frame, it comes back on the other side. I made a funtion the found the window's limits so it didn't have to be written out multiple times. It's the same script from fig 1.2