Refactor Example I.5
This one was a bit confusing because the output in Unity as I found it does not match the WebGL build shown on the site. We should discuss what we want out of this example. I refactored the code to look more like the example on the site but that version ran into a few problems that would need a good deal of refactoring to solve. I have a good idea of what those refactors would look like but we should discuss before implementing.

Capitalize method names

simplify FindWindowLimits()

refactor comments

rename classes and instances according to proposed guidelines